# docs/resultados/api_evidence/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThriveMind API",
    "description": "Ecosistema de bienestar holístico orquestado por IA",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/gamification/": {
      "get": {
        "tags": [
          "Gamificación"
        ],
        "summary": "Obtener puntos y hitos del usuario",
        "operationId": "get_gamification_api_v1_gamification__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/gamification/award": {
      "post": {
        "tags": [
          "Gamificación"
        ],
        "summary": "Otorgar puntos por acción",
        "operationId": "post_award_api_v1_gamification_award_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action"
            }
          },
          {
            "name": "referencia_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Referencia Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hue/profiles": {
      "get": {
        "tags": [
          "Iluminación HUE"
        ],
        "summary": "Listar perfiles predefinidos + custom",
        "operationId": "list_profiles_api_v1_hue_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/hue/profiles/custom": {
      "post": {
        "tags": [
          "Iluminación HUE"
        ],
        "summary": "Crear perfil personalizado",
        "operationId": "create_profile_api_v1_hue_profiles_custom_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HueProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/hue/profiles/{profile_id}": {
      "delete": {
        "tags": [
          "Iluminación HUE"
        ],
        "summary": "Eliminar perfil custom",
        "operationId": "delete_profile_api_v1_hue_profiles__profile_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HueProfileCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kelvin": {
            "type": "integer",
            "title": "Kelvin"
          },
          "brightness": {
            "type": "integer",
            "title": "Brightness"
          },
          "color_hex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Hex"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name",
          "kelvin",
          "brightness"
        ],
        "title": "HueProfileCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}